♻️ 코드 리팩토링 : StartSeedMoney Long으로 변경 , 빠른 매칭시 방 찾는 로직 변경

+++ backend/src/main/java/com/kmbbj/backend/matching/dto/CreateRoomDTO.java
@@ -15,7 +15,7 @@ public class CreateRoomDTO {
     private String title;
 
     // 시작 시드머니
-    private int startSeedMoney;
+    private Long startSeedMoney;
 
     // 마지막 라운드번호
     private int end;
